Module oak_loader::options

Helper functions to parse input arguments and create an instance of RuntimeConfiguration to feed into the oak_loader.

Structs

ConfigEntry

A specification of a configuration entry as human readable key and a path to a file whose contents constitutes the actual value.

Opt

Command line options for the Oak loader.

Functions

create_app_config

Parse application configuration into an instance of [ApplicationConfiguration]

create_grpc_config

Create the overall [oak_runtime::GrpcConfiguration] from the TLS certificate and private key files or using the Proxy Attestation Service.

create_http_config

Create the overall [oak_runtime::HttpConfiguration] from the TLS certificate and private key files.

create_permissions_config

Parse permissions configuration into an instance of [PermissionsConfiguration], if oak-unsafe is not enabled. Otherwise, use a default [PermissionsConfiguration].

create_runtime_config

Parse input options and create a RuntimeConfiguration.

create_secure_server_config

Create [oak_runtime::SecureServerConfiguration] containing optional TLS configurations for HTTP and gRPC server nodes.

create_sign_table

Create a signature table for Oak runtime. Returns an [SignatureTable] that maps each module hash to a vector of [SignatureBundle]. Returned signatures are not verified yet, they are supposed to be verified by the oak_runtime.

get_default_root_tls_certs

Gets the default root TLS certificates from the embedded byte array.

get_oidc_client_info

Parse OpenID Connect client configuration file into a [ClientInfo] .

get_root_tls_certificate_or_default

If oak-unsafe is enabled, reads root TLS certificate from the specified file into a byte array. Otherwise, loads the default root TLS certificate from the embedded byte array. Parses the byte array into a [Certificate], or returns an error if the byte array does not represent a valid PEM formatted certificate.

get_tls_identity

Creates Identity from the TLS certificate and private key files or using the Proxy Attestation Service.

parse_config_map

Parse a collection of configuration entries and return the contents of the corresponding files as a [ConfigMap].